[_config.styl] disable penguin background for the time being
The recent change (in site-design) to making the footer fixed to the
bottom of the viewport is a bit hacky because main is display: table,
and so main still has a fixed minimum height.  The penguin logo needs
to be at the bottom of the page, and I can't seem to get that working
with this particular flexbox setup.

The proper fix is to make main not be display: table.